requirements.txt
grpcio
protobuf
python-dotenv

grpcio-tools